pull out hive_api from db methods
